feat: add atuin and fzf-tab integrations
Adds two terminal power tools as first-class envi integrations:

- atuin: Warp-style searchable history popup on Up-arrow / Ctrl+R.
  Installs via Homebrew, auto-imports existing zsh history on first
  run, ships a compact inline-TUI default config that is only
  written when the user has none.

- fzf-tab: replaces zsh tab-completion menu with an fzf popup.
  Installed via Oh-My-Zsh custom plugin, loaded last in
  OHMYZSH_PLUGINS, configured with cd-folder preview via eza.

setup/_func_oh_my_zsh.sh plugin clones are now idempotent so the
setup can be re-run on existing installations.

+++ setup/_func_oh_my_zsh.sh
@@ -1,10 +1,21 @@
 #!/usr/bin/env bash
 
 function install_packages () {
-    curl -Lo /tmp/oh_my_install.sh https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh
-    sh /tmp/oh_my_install.sh --unattended
-    rm /tmp/oh_my_install.sh
+    if [ ! -d "${ZSH:-$HOME/.oh-my-zsh}" ]; then
+        curl -Lo /tmp/oh_my_install.sh https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh
+        sh /tmp/oh_my_install.sh --unattended
+        rm /tmp/oh_my_install.sh
+    fi
 
-    # Also install autosuggestions
-    git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
+    local plugins_dir="${ZSH_CUSTOM:-${ZSH:-$HOME/.oh-my-zsh}/custom}/plugins"
+
+    # zsh-autosuggestions: ghosted inline suggestions from history.
+    if [ ! -d "$plugins_dir/zsh-autosuggestions" ]; then
+        git clone https://github.com/zsh-users/zsh-autosuggestions "$plugins_dir/zsh-autosuggestions"
+    fi
+
+    # fzf-tab: fzf-powered popup for zsh tab-completion.
+    if [ ! -d "$plugins_dir/fzf-tab" ]; then
+        git clone https://github.com/Aloxaf/fzf-tab "$plugins_dir/fzf-tab"
+    fi
 }
